+++ R/nchar.R
@@ -80,7 +80,7 @@
 #'
 #'
 #' @seealso
-#' Related function(s): \code{\link{sprintf}}, \code{\link{substring}},
+#' Related function(s): \code{\link{sprintf}}, \code{\link{substr}},
 #'     \code{\link{strtrim}}
 #'
 #' @rdname nchar