+++ R/strtrim.R
@@ -73,7 +73,7 @@
 #' stringx::strtrim(x, 2)
 #'
 #' @seealso
-#' Related function(s): \code{\link{sprintf}}, \code{\link{substring}},
+#' Related function(s): \code{\link{sprintf}}, \code{\link{substr}},
 #'     \code{\link{nchar}}
 #'
 #' @rdname strtrim